+++ tests/PandoTests/Tests/Repositories/InMemoryRepositoryTests.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.IO;
 using FluentAssertions;
+using Pando;
 using Pando.Exceptions;
 using Pando.Repositories;
 using Pando.Repositories.Utils;
